requirements.txt
whisperx
torch
python-dotenv
pyyaml
